src/test/javascript/spec/app/entities/or-organization-unit/or-organization-unit.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { OrOrganizationUnitService } from 'app/entities/or-organization-unit/or-organization-unit.service';
import { IOrOrganizationUnit, OrOrganizationUnit } from 'app/shared/model/or-organization-unit.model';

describe('Service Tests', () => {
  describe('OrOrganizationUnit Service', () => {
    let injector: TestBed;
    let service: OrOrganizationUnitService;
    let httpMock: HttpTestingController;
    let elemDefault: IOrOrganizationUnit;
    let expectedResult: IOrOrganizationUnit | IOrOrganizationUnit[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(OrOrganizationUnitService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new OrOrganizationUnit(0, 'AAAAAAA', 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a OrOrganizationUnit', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new OrOrganizationUnit()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a OrOrganizationUnit', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            parentId: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of OrOrganizationUnit', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            parentId: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a OrOrganizationUnit', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
